+++ src/ReqRest.Client.Tests/DtoMocks.cs
@@ -1,0 +1,16 @@
+﻿namespace ReqRest.Client.Tests
+{
+
+    // For the generic ApiRequest{...} classes, we need some kind of type parameters in the tests.
+    // Anything is fine, so let's just declare a common set of classes.
+
+    public class Dto1 { }
+    public class Dto2 { }
+    public class Dto3 { }
+    public class Dto4 { }
+    public class Dto5 { }
+    public class Dto6 { }
+    public class Dto7 { }
+    public class Dto8 { }
+
+}